backend/requirements.txt
fastapi>=0.121.3
uvicorn>=0.38.0
python-dotenv>=1.2.1
requests>=2.32.5

# Database & Data Processing

pandas>=2.3.3
chromadb>=1.3.5
pydantic>=2.12.4

# LangChain Ecosystem

langchain>=1.1.0
langchain-community>=0.4.1
langchain-core>=1.1.0
langchain-google-genai>=3.2.0
langchain-huggingface>=1.0.1
langchain-chroma>=1.0.0
langchain-groq>=1.1.0
langchain-text-splitters>=1.0.0

# Document Loaders

pypdf>=6.4.0

# Database Extraction (SSH Tunneling)

psycopg2-binary>=2.9.11
paramiko>=3.0.0
sshtunnel>=0.4.0